pyproject.toml
[tool.poetry]
name = "user_management"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.13.1"
pytest = "^8.1.1"
uvicorn = {extras = ["standart"], version = "^0.29.0"}
pydantic = {extras = ["email"], version = "^2.6.4"}
annotated-types = "^0.6.0"
fastapi = "^0.110.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
gunicorn = "^21.2.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"
pytz = "^2024.1"
redis = "^5.0.3"
requests = "^2.31.0"
aio-pika = "^9.4.1"
aioredis = "^2.0.1"
httpx = {extras = ["http2"], version = "^0.27.0"}
pytest-asyncio = "^0.23.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  ".", "scr",
]
asyncio_mode="auto"
